+++ api/client.js
@@ -1,13 +1,15 @@
-// SPDX-FileCopyrightText: 2022 Sveriges Television AB
+// SPDX-FileCopyrightText: 2026 Axel Boberg
 //
 // SPDX-License-Identifier: MIT
 
+const environment = require('./shared/environment')
+
 ;(function () {
-  if (module.parent) {
+  if (environment.isNode()) {
     require('./node/client')
     return
   }
-  if (typeof window !== 'undefined') {
+  if (environment.isBrowser()) {
     require('./browser/client')
   }
 })()