+++ api/shared/environment.js
@@ -1,0 +1,9 @@
+function isNode () {
+  return !!module.parent
+}
+exports.isNode = isNode
+
+function isBrowser () {
+  return typeof window !== 'undefined'
+}
+exports.isBrowser = isBrowser