+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Human/AI readable descriptors for types
 - Free wheel for LTC inputs
 - Reference items now show a warning if they're missing a target id
+- Keyboard triggers
 ### Fixed
 - An issue where invalid types could crash the rundown widget
 - An issue where 0 (or falsy values) could not be used as ids for options in preferences of the select type
@@ -14,6 +15,7 @@
 - Settings no longer collide with other modals
 ### Changed
 - Keyboard shortcuts now use hotkeys-js for better stability
+- The item.apply event has been changed to item.change
 
 ## 1.0.0-beta.9
 ### Added